Rename mesh_client_mock to orders
